Rename the junit5 folder to junit4 and adjust package declarations.
