+++ jvm/undertest-junit4/src/test/kotlin/undertest/junit4/SelfieWriteOnce.kt
@@ -1,4 +1,4 @@
-package undertest.junit5
+package undertest.junit4
 
 import com.diffplug.selfie.junit5.SelfieSettingsAPI
