+++ jvm/undertest-junit4/src/test/kotlin/undertest/junit4/UT_StringLiteralsKotlinTest.kt
@@ -1,4 +1,4 @@
-package undertest.junit5
+package undertest.junit4
 
 import com.diffplug.selfie.Selfie.expectSelfie
 import org.junit.jupiter.api.Test
